+++ controller/rq_definition.ml
@@ -12,7 +12,7 @@ let request ~(doc : Fleche.Doc.t) ~point =
       match CString.Map.find_opt id_at_point toc with
       | Some range ->
         let uri = doc.uri in
-        Lsp.JFleche.Location.({ uri; range } |> to_yojson)
+        Lsp.Core.Location.({ uri; range } |> to_yojson)
       | None -> `Null)
     `Null
     (Rq_common.get_id_at_point ~doc ~point)