+++ editor/code/src/config.ts
@@ -3,7 +3,6 @@ import { ExtensionContext, workspace } from "vscode";
 export interface CoqLspServerConfig {
   client_version: string;
   eager_diagnostics: boolean;
-  ok_diagnostics: boolean;
   goal_after_tactic: boolean;
   show_coq_info_messages: boolean;
   show_notices_as_diagnostics: boolean;
@@ -22,7 +21,6 @@ export namespace CoqLspServerConfig {
     return {
       client_version: client_version,
       eager_diagnostics: wsConfig.eager_diagnostics,
-      ok_diagnostics: wsConfig.ok_diagnostics,
       goal_after_tactic: wsConfig.goal_after_tactic,
       show_coq_info_messages: wsConfig.show_coq_info_messages,
       show_notices_as_diagnostics: wsConfig.show_notices_as_diagnostics,