+++ lsp/base.ml
@@ -23,7 +23,9 @@ module J = Yojson.Safe
 module U = Yojson.Safe.Util
 
 let string_field name dict = U.to_string List.(assoc name dict)
-let dict_field name dict = U.to_assoc (List.assoc name dict)
+
+let odict_field ~default name dict =
+  Option.cata U.to_assoc default (List.assoc_opt name dict)
 
 module Message = struct
   type t =
@@ -42,7 +44,7 @@ module Message = struct
     try
       let dict = U.to_assoc msg in
       let method_ = string_field "method" dict in
-      let params = dict_field "params" dict in
+      let params = odict_field ~default:[] "params" dict in
       (match List.assoc_opt "id" dict with
       | None -> Notification { method_; params }
       | Some id ->
@@ -77,3 +79,61 @@ let mk_notification ~method_ ~params =
     ; ("method", `String method_)
     ; ("params", params)
     ]
+
+module ProgressToken : sig
+  type t =
+    | String of string
+    | Int of int
+  [@@deriving yojson]
+end = struct
+  type t =
+    | String of string
+    | Int of int
+
+  let of_yojson x =
+    match x with
+    | `String s -> Result.ok (String s)
+    | `Int i -> Result.ok (Int i)
+    | _ -> Result.error "failure to parse ProgressToken.t"
+
+  let to_yojson = function
+    | String s -> `String s
+    | Int i -> `Int i
+end
+
+module ProgressParams = struct
+  type 'a t =
+    { token : ProgressToken.t
+    ; value : 'a
+    }
+  [@@deriving yojson]
+end
+
+let mk_progress ~token ~value f =
+  let params = ProgressParams.(to_yojson f { token; value }) in
+  mk_notification ~method_:"$/progress" ~params
+
+module WorkDoneProgressBegin = struct
+  type t =
+    { kind : string
+    ; title : string
+    ; cancellable : bool option [@None]
+    ; message : string option [@None]
+    ; percentage : int option [@None]
+    }
+  [@@deriving to_yojson]
+end
+
+module WorkDoneProgressReport = struct
+  type t =
+    { kind : string
+    ; cancellable : bool option [@None]
+    ; message : string option [@None]
+    ; percentage : int option [@None]
+    }
+  [@@deriving to_yojson]
+end
+
+module WorkDoneProgressEnd = struct
+  type t = { kind : string } [@@deriving to_yojson]
+end