+++ lsp/io.mli
@@ -18,10 +18,10 @@
 (** JSON-RPC input/output *)
 
 (** Read a JSON-RPC request from channel *)
-val read_raw_request : in_channel -> Yojson.Safe.t
+val read_raw_request : in_channel -> Yojson.Safe.t option
 
-(** *)
-val read_request : in_channel -> Base.Message.t
+(** [None] signals [EOF] *)
+val read_request : in_channel -> Base.Message.t option
 
 exception ReadError of string